Format Gen4 end-to-end tests
