+++ go/test/endtoend/vtgate/queries/misc/schema.sql
@@ -1,0 +1,5 @@
+create table t1(
+                   id1 bigint,
+                   id2 bigint,
+                   primary key(id1)
+) Engine=InnoDB;
