+++ go/test/endtoend/vtgate/queries/normalize/main_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package normalize
 
 import (
+	_ "embed"
 	"flag"
 	"os"
 	"testing"
@@ -30,26 +31,11 @@ var (
 	vtParams        mysql.ConnParams
 	keyspaceName    = "ks_normalize"
 	cell            = "test_normalize"
-	schemaSQL       = `
-create table t1(
-  id bigint unsigned not null,
-  charcol char(10),
-  vcharcol varchar(50),
-  bincol binary(50),
-  varbincol varbinary(50),
-  floatcol float,
-  deccol decimal(5,2),
-  bitcol bit,
-  datecol date,
-  enumcol enum('small', 'medium', 'large'),
-  setcol set('a', 'b', 'c'),
-  jsoncol json,
-  geocol geometry,
-  binvalcol varbinary(50),
-  binnumcol varbinary(50),
-  primary key(id)
-) Engine=InnoDB;
-`
+
+	//go:embed schema.sql
+	schemaSQL string
+	//go:embed vschema.json
+	_ string
 )
 
 func TestMain(m *testing.M) {
